# integration-framework/moirai-zeus/src/main/resources/static/swagger/apis/openapi.yaml
openapi: 3.0.0
info:
  title: Moirai Zeus API
  description: This is the API for the Zeus(Ta3) system's access to the subprojects (e.g. Clotho),
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /services/list:
    get:
      tags:
        - Services
      summary: Get services created by user accounts.
      description: get service by parameter list
      operationId: listServices
      parameters:
        - name: name
          in: query
          description: name value for service
          schema:
            type: string
            example: Service Name
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/service'
        404:
          description: Not found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Invalid credentials.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services:
    put:
      tags:
        - Services
      summary: Updates an instance of a service
      description: updates service
      operationId: updateServiceInstance
      requestBody:
        description: updates service with updated details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/service'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
    delete:
      tags:
        - Services
      summary: Terminates an instance of a service
      description: termninates service
      operationId: terminateServiceInstance
      requestBody:
        description: Terminates service with the selected parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/service'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: An existing provision for this user already exists
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services/clotho:
    post:
      tags:
        - Services
      summary: Provisions an instance of Clotho with a user selected database
      description: Initialize Clotho with a user selected database
      operationId: provisionClotho
      requestBody:
        description: Launches Terraform with the selected parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/clothoServiceDto'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: An existing provision for this user already exists
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
    delete:
      tags:
        - Services
      summary: Terminates an instance of Clotho with a user selected database
      description: termninates Clotho with a user selected database
      operationId: terminateClothoInstance
      requestBody:
        description: Terminates clotho with the selected parameters
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/service'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: An existing provision for this user already exists
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services/docker/registry/password/reset:
    post:
      tags:
        - Docker
      summary: Resets the Docker registry password for the logged in user
      description: Reset the Docker registry password
      operationId: resetDockerRegistryPassword
      requestBody:
        description: Password to reset password to
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services/docker/registry/refresh:
    post:
      tags:
        - Docker
      summary: Forces the zeus to update the docker registry
      operationId: refreshDockerRegistry
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        404:
          description: Not Found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services/docker/registry/restore:
    post:
      tags:
        - Docker
      summary: Recreate the docker registries basd on info in the DB.
      operationId: restoreDockerRegistry
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services/docker/registry:
    get:
      tags:
        - Docker
      summary: Retrieves the loged in user's Docker registry.
      description: retrieve docker registry
      operationId: getDockerRegistry
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/dockerRegistry'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        404:
          description: Not Found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
    post:
      tags:
        - Docker
      summary: Provisions a Private Docker registry for a performer
      description: Provisions a Private Docker registry for a performer
      operationId: createDockerRegistry
      requestBody:
        description: Connects to a Flask service that creates a private Docker registry.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/stringRequest'
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        409:
          description: An existing provision for this user already exists
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
    delete:
      tags:
        - Docker
      summary: Destroys Docker uploads
      description: Destroys Docker upload
      operationId: destroyDockerUploads
      parameters:
        - name: id
          in: query
          description: DockerUpload id of DockerUpload to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        400:
          description: Bad request. Invalid input, object invalid
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        401:
          description: Unauthorized.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
  /services/docker/registry/list:
    get:
      tags:
        - Docker
      summary: Retrieves a docker registry for the username.
      description: get docker registry for user
      operationId: getDockerRegistryForUser
      parameters:
        - name: username
          in: query
          description: UserName
          schema:
            type: string
            example: Performer Name
      responses:
        200:
          description: OK
          content:
            'application/json':
              schema:
                items:
                  $ref: '#/components/schemas/dockerRegistry'
        401:
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
        404:
          description: Not Found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/stringResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    dockerRegistry:
      type: object
      x-jpaEntity: true
      required:
        - owner
        - endpoint
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
          x-read-only: true
        owner:
          type: string
        endpoint:
          type: string
        password:
          type: string
        dockerimagelist:
          type: array
          items:
            $ref: '#/components/schemas/dockerUpload'
          x-lazyOneToMany: true
          x-cascade-all: true
    dockerUpload:
      type: object
      x-jpaEntity: true
      required:
        - registry
        - repo
        - tag
        - dockerimagelocation
        - lastModified
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
          x-read-only: true
        registry:
          type: string
          example: "Performer 1"
        repo:
          type: string
          example: "Image 1"
        tag:
          type: string
        lastModified:
          type: string
          format: date-time
          example: "2020-01-01"
        digest:
          type: string
        size:
          type: integer
        dockerimagelocation:
          type: string
          example: "s3://kairos-service-artifacts/connectors/next_century/service_a/connector.py"
...
    clothoServiceDto:
      type: object
      required:
        - name
        - details
        - databaseType
      properties:
        name:
          type: string
          example: service10
        databaseType:
          type: string
          enum:
            - Neo4j
            - JanusGraph
            - OrientDB
        details:
          type: string
          example: service that will help me perform tasks A, B, and C
...
    stringResponse:
      type: object
      properties:
        value:
          type: string
          example: "Success"
    stringRequest:
      type: object
      properties:
        value:
          type: string
          example: "Sample_text"
...
    service:
      type: object
      x-jpaEntity: true
      required:
        - name
      properties:
        id:
          type: string
          example:
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "service1"
        type:
          type: string
          example: "clotho"
        subtype:
          type: string
          example: "Neo4j"
        status:
          type: string
        dateReceived:
          type: string
          format: date
          example: "2020-01-01"
        teamName:
          type: string
        access:
          type: string
          example: "Private"
        details:
          type: string
        error:
          type: string
          example: "Error Msg"
          x-jpaTransient: true
        awsInstances:
          type: array
          items:
            $ref: '#/components/schemas/serviceAwsInstance'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_aws_instance",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_aws_instance_id") }
          x-cascade-all: true
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/serviceEndpoint'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_endpoint",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_endpoint_id") }
          x-cascade-all: true
        downloads:
          type: array
          items:
            $ref: '#/components/schemas/serviceDownload'
          x-lazyOneToMany: true
          x-jsonIgnoreProperty:
            propertyKey: missions
          x-joinTable:
            joinValue: name = "service_service_download",
              joinColumns = { @JoinColumn(name = "service_id") },
              inverseJoinColumns = { @JoinColumn(name = "service_download_id") }
          x-cascade-all: true
    serviceAwsInstance:
      type: object
      x-jpaEntity: true
      required:
        - instanceId
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        instanceId:
          type: string
          example: "i-011d626d652684d5a4"
    serviceEndpoint:
      type: object
      x-jpaEntity: true
      required:
        - uri
        - name
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "Clotho Endpoint"
        uri:
          type: string
          example: "ec2-18-232-154-122.compute-1.amazonaws.com:8008"
    serviceDownload:
      type: object
      x-jpaEntity: true
      required:
        - uri
        - name
      properties:
        id:
          type: string
          x-jpaIdField: true
          x-generatedId: true
        name:
          type: string
          example: "Python Connector"
        uri:
          type: string
          example: "s3://kairos-service-artifacts/connectors/next_century/service_a/connector.py"
...